+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/Composer/Audio/AudioRecorderComposerPresenterImpl.swift
@@ -39,7 +39,7 @@ class AudioRecorderComposerPresenterImpl: AudioRecorderComposerPresenter {
                     self.delegate?.audioRecorderComposerPresenterCanNotStartRecording(self)
                 }
             case let .failure(error):
-                self.logger.error(message: "Unable to start recording audio message: \(error)")
+                self.logger.warning(message: "Unable to start recording audio message", extra: ["reason": error])
                 self.delegate?.audioRecorderComposerPresenterCanNotStartRecording(self)
             }
         }